pyproject.toml
[tool.poetry]
name = "ad-bidding-agent"
version = "0.1.0"
description = "LangGraph Ad Bidding Recommendation System"
authors = []
readme = "README.md"
packages = [{include = "agent"}, {include = "api"}, {include = "data"}, {include = "eval"}]

[tool.poetry.dependencies]
python = "^3.11"

# Agent / LLM
langgraph = "^0.2"
langchain = "^0.3"
langchain-aws = "^0.2"
langchain-community = "^0.3"
langsmith = "^0.1"

# API
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.30"}
python-jose = {extras = ["cryptography"], version = "^3.3"}
httpx = "^0.27"

# Data / Warehouse
snowflake-connector-python = {extras = ["pandas"], version = "^3.10"}
dbt-snowflake = "^1.8"
redis = "^5.0"
pydantic = "^2.8"
pydantic-settings = "^2.4"

# Streaming
confluent-kafka = "^2.5"
fastavro = "^1.9"

# MLOps
mlflow = "^2.15"
boto3 = "^1.35"

# Guardrails
nemoguardrails = "^0.10"

# Observability
prometheus-client = "^0.21"
opentelemetry-sdk = "^1.27"
opentelemetry-instrumentation-fastapi = "^0.48"

# Utilities
python-dotenv = "^1.0"
structlog = "^24.4"
tenacity = "^9.0"
numpy = "^1.26"
pandas = "^2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pytest-asyncio = "^0.24"
pytest-cov = "^5.0"
ruff = "^0.6"
mypy = "^1.11"
pre-commit = "^3.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
